[chore]: moved xtasks entirely to CLI
