--- .cargo/config.toml
@@ -1,2 +1,0 @@
-[alias]
-xtask = "run --manifest-path ./xtask/Cargo.toml --"
