+++ .gitignore
@@ -4,7 +4,7 @@ debug/
 target/
 target_ra/
 
-# Generated by `cargo xtask build`
+# Generated by `cargo run build`
 components/
 
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
