+++ Cargo.toml
@@ -1,7 +1,6 @@
 [workspace]
 members = [
   # Tooling
-  "xtask",
   "protos",
   "cli",
   "development_server",
