--- xtask/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "xtask"
-description = "Custom Mycelia tasks"
-version = "0.1.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-repository = "https://github.com/MyceliaNetwork/mycelia"
-authors = ["Daniel Terwiel <dani+mycelia@departure.dev>"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-thiserror = { workspace = true }
-dialoguer = "0.11.0"
-env_logger = { workspace = true }
-log = { workspace = true }
-tokio = { version = "1.32.0", features = ["full"] }
-trim-margin = "0.1.0"
-semver = { workspace = true }
-cargo-workspaces = { version = "0.2.44" }
-cargo_metadata = { version = "0.18.0" }
-octocrab = { version = "0.31.1" }
-chrono = "0.4.31"
